requirements.txt

cmake==3.18.4.post1
dlib==19.21.1
face-recognition==1.3.0
numpy==1.20.1
opencv-contrib-python==4.5.1.48
Pillow==8.1.2
